+++ src/components/catalog/label/index.js
